refactor: added infra layer
