+++ src/modules/cars/infra/repositories/CategoriesRepository.ts
@@ -1,6 +1,10 @@
-import { getRepository, Repository } from "typeorm"
-import { Category } from "../../entities/Category"
-import { ICategoriesRepository, ICreateCategoryDTO } from "../ICategoriesRepository"
+import { getRepository, Repository } from 'typeorm'
+
+import { Category } from '../typeorm/entities/Category'
+import {
+  ICategoriesRepository,
+  ICreateCategoryDTO
+} from '../../repositories/ICategoriesRepository'
 
 class CategoriesRepository implements ICategoriesRepository {
