+++ src/modules/cars/infra/repositories/SpecificationsRepository.ts
@@ -1,7 +1,10 @@
-import { getRepository, Repository } from "typeorm";
+import { getRepository, Repository } from 'typeorm'
 
-import { Specification } from "../../entities/Specification";
-import { ISpecificationsRepository, ICreateSpecificationDTO } from "../ISpecificationsRepository";
+import { Specification } from '../typeorm/entities/Specification'
+import {
+  ICreateSpecificationDTO,
+  ISpecificationsRepository
+} from '../../repositories/ISpecificationsRepository'
 
 class SpecificationsRepository implements ISpecificationsRepository {
   private repository: Repository<Specification>
